+++ deepclone.stub.php
@@ -11,9 +11,15 @@ namespace DeepClone {
 }
 
 namespace {
+    /**
+     * @var int
+     * @cvalue DEEPCLONE_HYDRATE_CALL_HOOKS
+     */
+    const DEEPCLONE_HYDRATE_CALL_HOOKS = UNKNOWN;
+
     function deepclone_to_array(mixed $value, ?array $allowed_classes = null): array {}
 
     function deepclone_from_array(array $data, ?array $allowed_classes = null): mixed {}
 
-    function deepclone_hydrate(object|string $object_or_class, array $scoped_vars = [], array $mangled_vars = []): object {}
+    function deepclone_hydrate(object|string $object_or_class, array $scoped_vars = [], array $mangled_vars = [], int $flags = 0): object {}
 }